Using GeoLocation Object, Object Destructuring, and Template Literals
Checked to ensure navigator object exists to the browser. Used its callback parameters to create two functions. A parameterless function in the case the position could not be found and a function that accepted the found position as a parameter. The function then takes the latitude and longitude from the coords object using Destructuring. Then we use those variables in a logged template literal that represents a valid link to google maps using the same values.