Completed Fundamental Class Architecture
Defined a Class of "App" that contains two private member variables, #map and #mapEvent, that can be referenced in the class functions with the use of the .bind.(this) keyword so the "this" keyword references the private member instances of #map and #mapEvent instead of referencing the function or objects they're attached to.

Proceeding to define the Workout Class and the Derived Classes "Running" and "Cycling" for more organized definition and handling of a Running list element and a Cycling list element as they're appended.