Defined Workout Object Class, Running and Cycling Derived Classes
Defined parameterized classes for Running and Cycling workouts that inherit parameters and information from the Workout parent class.

Current ID implementation not ready for production as it uses Date in a asynchronous multi-user application. Implementing this kind of unique ID methodology will result in the creation of multiple objects with the same ID number by nature of the possibility of users creating objects at exactly the same time.

Moving forward with rendering Workout data to marker pop-ups and populating the Workout List.