App V1 Feature Complete
All intended features for the application have been implemented.

Needed to remove clicks public function because it dependent on style classes. Stringifying a JSON object and parsing it back from JSON to a string destroys the prototype chain; removing any former classes or ids the object previously had. They can be restored using identifiers and iteration, but was outside of the scope of this project.

Included console-activated "reset" method in app to clear the browser local storage of object data after use. Could implement CRUD operations in the UI to make more user friendly. Was outside of the scope of this project.

I learned a great deal about template literals, destructuring values from objects, binding the "this" keyword and when to do so for class methods that may be attached to other objects, event listeners, or callback functions, and using insertAdjacentHTML to concatenate HTML values with string literals into UI elements directly onto the DOM.

Was a fascinating project and may return to it in the future to implement recommended changes/feature additions.